+++ WidgetExtension/CourseDateOverviewWidget.swift
@@ -6,62 +6,6 @@
 import SwiftUI
 import WidgetKit
 
-
-import Common
-
-struct CourseDateOverviewView: View {
-
-    var courseDateOverview: CourseDateOverviewViewModel
-
-    var body: some View {
-        VStack {
-            Text("Course Date Overview")
-                .font(.headline)
-                .frame(maxWidth: .infinity, alignment: .leading)
-            Divider()
-            Spacer()
-            HStack {
-                Text("Today")
-                Spacer()
-                Text("\(courseDateOverview.todayCount)")
-                    .font(.callout)
-                    .padding([.leading, .trailing])
-                    .padding([.top, .bottom], 2)
-                    .foregroundColor(.white)
-                    .background(Color.blue)
-                    .clipShape(Capsule())
-            }
-            Spacer()
-            HStack {
-                Text("Next 7 Days")
-                Spacer()
-                Text("\(courseDateOverview.nextCount)")
-                    .font(.callout)
-                    .padding([.leading, .trailing])
-                    .padding([.top, .bottom], 2)
-                    .foregroundColor(.white)
-                    .background(Color.blue)
-                    .clipShape(Capsule())
-            }
-            Spacer()
-            HStack {
-                Text("All")
-                Spacer()
-                Text("\(courseDateOverview.allCount)")
-                    .font(.callout)
-                    .padding([.leading, .trailing])
-                    .padding([.top, .bottom], 2)
-                    .foregroundColor(.white)
-                    .background(Color.blue)
-                    .clipShape(Capsule())
-            }
-            Spacer()
-        }
-    }
-
-}
-
-
 struct CourseDateOverviewWidgetEntryView : View {
     var entry: CourseDateOverviewWidgetProvider.Entry
 
@@ -90,20 +34,6 @@ struct CourseDateOverviewWidget: Widget {
         .configurationDisplayName("Course Date Overview")
         .description("This is an example widget.")
         .supportedFamilies([.systemSmall])
-    }
-
-}
-
-struct CourseDateOverviewWidget_Previews: PreviewProvider {
-
-    static var exampleCourseDateOverview: CourseDateOverviewViewModel {
-        CourseDateOverviewViewModel(todayCount: 1, nextCount: 2, allCount: 4)
-    }
-
-    static var previews: some View {
-        CourseDateOverviewWidgetEntryView(entry: CourseDateOverviewWidgetEntry(courseDateOverview: exampleCourseDateOverview, userIsLoggedIn: true))
-                .previewContext(WidgetPreviewContext(family: .systemSmall))
-                .previewDevice(PreviewDevice(rawValue: "iPhone SE"))
     }
 
 }
