+++ WidgetExtension/Views/EmptyContentView.swift
@@ -4,6 +4,7 @@
 //
 
 import SwiftUI
+import WidgetKit
 
 struct EmptyContentView: View {
 
@@ -16,6 +17,24 @@ struct EmptyContentView: View {
                 .multilineTextAlignment(.center)
                 .font(.system(.footnote))
                 .foregroundColor(Color.secondary)
+        }
+    }
+
+}
+
+struct EmptyContentView_Previews: PreviewProvider {
+
+    static var previews: some View {
+        Group {
+            EmptyContentView()
+                .padding()
+                .previewContext(WidgetPreviewContext(family: .systemSmall))
+                .previewDevice(PreviewDevice(rawValue: "iPhone SE"))
+
+            EmptyContentView()
+                .padding()
+                .previewContext(WidgetPreviewContext(family: .systemMedium))
+                .previewDevice(PreviewDevice(rawValue: "iPhone SE"))
         }
     }
